Minor changes
Combined PortalBuilder classes from parse and database packages for
less confusion.
Added a status dialog for authentication and email search
Updated code style settings
