+++ app/src/main/java/com/einzig/ipst2/database/PortalAcceptedBuilder.java
@@ -1,5 +1,4 @@
 /******************************************************************************
- *                                                                            *
  * Copyright 2017 Steven Foskett, Jimmy Ho, Ryan Porterfield                  *
  * Permission is hereby granted, free of charge, to any person obtaining a    *
  * copy of this software and associated documentation files (the "Software"), *
@@ -18,7 +17,6 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING    *
  * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER        *
  * DEALINGS IN THE SOFTWARE.                                                  *
- *                                                                            *
  ******************************************************************************/
 
 package com.einzig.ipst2.database;
@@ -27,23 +25,35 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 
 import com.einzig.ipst2.portal.PortalAccepted;
+import com.einzig.ipst2.portal.PortalSubmission;
 
-import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 /**
  * @author Ryan Porterfield
  * @since 2017-05-19
  */
 
-final class PortalAcceptedBuilder extends PortalBuilder<PortalAccepted> {
-
+public final class PortalAcceptedBuilder extends PortalBuilder<PortalAccepted> {
     /**
-     * @param dateFormatter date format that MySQL uses to store DATETIME objects
      * @param db reference to a SQLite database to run queries on
      */
-    PortalAcceptedBuilder(SimpleDateFormat dateFormatter, SQLiteDatabase db) {
-        super(dateFormatter, db, DatabaseInterface.TABLE_ACCEPTED);
+    public PortalAcceptedBuilder(SQLiteDatabase db) {
+        super(db, DatabaseInterface.TABLE_ACCEPTED);
+    }
+
+    @Override
+    public PortalAccepted build(String name, Date dateResponded, String message) {
+        String pictureURL = parsePictureURL(message);
+        Date dateSubmitted = dateResponded;
+        PortalSubmission submission = findPortal(pictureURL);
+        if (submission != null)
+            dateSubmitted = submission.getDateSubmitted();
+        String address = parseLiveAddress(message);
+        String intelLink = parseIntelLink(message);
+        return new PortalAccepted(name, dateSubmitted, pictureURL, dateResponded, address, intelLink);
     }
 
     /**
@@ -52,7 +62,7 @@ final class PortalAcceptedBuilder extends PortalBuilder<PortalAccepted> {
      * @return a PortalAccepted representation of a portal in the database.
      */
     @Override
-    PortalAccepted createPortal(Cursor cursor) {
+    PortalAccepted build(Cursor cursor) {
         String name, pictureURL, location, intelLink;
         Date submitted, responded;
         name = cursor.getString(0);
@@ -62,5 +72,42 @@ final class PortalAcceptedBuilder extends PortalBuilder<PortalAccepted> {
         location = cursor.getString(4);
         intelLink = cursor.getString(5);
         return new PortalAccepted(name, submitted, pictureURL, responded, location, intelLink);
+    }
+
+    /**
+     * Parse the intel link for an accepted portal from the email.
+     *
+     * @param messageString The body of the email as a String for parsing.
+     * @return Ingress Intel map link for the portal.
+     */
+    private String parseIntelLink(String messageString) {
+        String intelLinkURL;
+        Pattern p = Pattern.compile("href=\"(.*?)\"");
+        Matcher m = p.matcher(messageString);
+        if (m.find()) {
+            intelLinkURL = m.group(1);
+        }
+        else
+            intelLinkURL = "N/A";
+        return intelLinkURL;
+    }
+
+    /**
+     * Parse the address of an accepted portal from the email.
+     *
+     * @param messageString The body of the email as a String for parsing.
+     * @return address of the portal.
+     */
+    private String parseLiveAddress(String messageString) {
+        String liveAddress;
+        Pattern titleFinder = Pattern.compile("<a[^>]*>(.*?)</a>",
+                Pattern.DOTALL | Pattern.CASE_INSENSITIVE);
+        Matcher regexMatcher = titleFinder.matcher(messageString);
+        if (regexMatcher.find()) {
+            liveAddress = regexMatcher.group(1);
+        }
+        else
+            liveAddress = "N/A";
+        return liveAddress;
     }
 }
