+++ app/src/main/java/com/einzig/ipst2/database/PortalRejectedBuilder.java
@@ -1,5 +1,4 @@
 /******************************************************************************
- *                                                                            *
  * Copyright 2017 Steven Foskett, Jimmy Ho, Ryan Porterfield                  *
  * Permission is hereby granted, free of charge, to any person obtaining a    *
  * copy of this software and associated documentation files (the "Software"), *
@@ -18,7 +17,6 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING    *
  * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER        *
  * DEALINGS IN THE SOFTWARE.                                                  *
- *                                                                            *
  ******************************************************************************/
 
 package com.einzig.ipst2.database;
@@ -27,8 +25,8 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 
 import com.einzig.ipst2.portal.PortalRejected;
+import com.einzig.ipst2.portal.PortalSubmission;
 
-import java.text.SimpleDateFormat;
 import java.util.Date;
 
 /**
@@ -36,13 +34,12 @@ import java.util.Date;
  * @since 2017-05-19
  */
 
-final class PortalRejectedBuilder extends PortalBuilder<PortalRejected> {
+public final class PortalRejectedBuilder extends PortalBuilder<PortalRejected> {
     /**
-     * @param dateFormatter date format that MySQL uses to store DATETIME objects
      * @param db reference to a SQLite database to run queries on
      */
-    PortalRejectedBuilder(SimpleDateFormat dateFormatter, SQLiteDatabase db) {
-        super(dateFormatter, db, DatabaseInterface.TABLE_REJECTED);
+    public PortalRejectedBuilder(SQLiteDatabase db) {
+        super(db, DatabaseInterface.TABLE_REJECTED);
     }
 
     /**
@@ -51,7 +48,7 @@ final class PortalRejectedBuilder extends PortalBuilder<PortalRejected> {
      * @return a PortalRejected representation of a portal in the database.
      */
     @Override
-    PortalRejected createPortal(Cursor cursor) {
+    PortalRejected build(Cursor cursor) {
         String name, pictureURL, reason;
         Date submitted, responded;
         name = cursor.getString(0);
@@ -60,5 +57,36 @@ final class PortalRejectedBuilder extends PortalBuilder<PortalRejected> {
         responded = parseDate(cursor.getString(3));
         reason = cursor.getString(4);
         return new PortalRejected(name, submitted, pictureURL, responded, reason);
+    }
+
+    @Override
+    public PortalRejected build(String name, Date dateResponded, String message) {
+        String pictureURL = parsePictureURL(message);
+        Date dateSubmitted = dateResponded;
+        PortalSubmission submission = findPortal(pictureURL);
+        if (submission != null)
+            dateSubmitted = submission.getDateSubmitted();
+        String rejectionReason = parseRejectionReason(message);
+        return new PortalRejected(name, dateSubmitted, pictureURL, dateResponded, rejectionReason);
+    }
+
+    /**
+     * Parse the reason a portal was rejected.
+     * @param messageString The body of the email as a String for parsing.
+     * @return the reason the portal was rejected.
+     */
+    private String parseRejectionReason(String messageString) {
+        String rejectionReason = "N/A";
+        if (messageString.contains("does not meet the criteria"))
+            rejectionReason = "Does not meet portal criteria";
+        if (messageString.contains("duplicate"))
+            rejectionReason = "Duplicate of another portal";
+        if (messageString.contains("too close")) {
+            if(rejectionReason.equalsIgnoreCase("N/A"))
+                rejectionReason = "Too Close to another portal";
+            else
+                rejectionReason = rejectionReason + " or too close to another portal";
+        }
+        return rejectionReason;
     }
 }
