+++ app/src/main/java/com/einzig/ipst2/parse/GetMailTask.java
@@ -34,18 +34,11 @@ import com.einzig.ipst2.DialogHelper;
 import com.einzig.ipst2.R;
 import com.einzig.ipst2.activities.MainActivity;
 import com.einzig.ipst2.oauth.OAuth2Authenticator;
-import com.google.android.gms.auth.GoogleAuthException;
-import com.google.android.gms.auth.GoogleAuthUtil;
-import com.google.android.gms.auth.GooglePlayServicesAvailabilityException;
 import com.sun.mail.imap.IMAPStore;
 
-import java.io.IOException;
-import java.text.DateFormat;
 import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
-import java.util.Locale;
 
 import javax.mail.FetchProfile;
 import javax.mail.Folder;
@@ -60,31 +53,27 @@ import javax.mail.search.ReceivedDateTerm;
 import javax.mail.search.SearchTerm;
 import javax.mail.search.SubjectTerm;
 
+import static com.einzig.ipst2.database.DatabaseInterface.dateFormatter;
+
 /**
  * @author Ryan Porterfield
  * @since 2017-05-28
  */
-
 public class GetMailTask extends AsyncTask<Void, Void, MailBundle> {
+    /** GMail account used for Ingress */
     final private Account account;
+    /** Parent activity of this task */
     final private Activity activity;
-    /**
-     * Format for parsing and printing dates
-     */
-    final private DateFormat dateFormat;
-    /**
-     * App preferences
-     */
+    /** App preferences */
     final private SharedPreferences preferences;
+    /** OAuth token */
     final private String token;
 
     public GetMailTask(Activity activity, Account account, String token) {
         this.account = account;
         this.activity = activity;
-        this.dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm", Locale.US);
         this.preferences = activity.getPreferences(MainActivity.MODE_PRIVATE);
         this.token = token;
-        Log.d(MainActivity.TAG, "GetMailTask:token -> " + token);
     }
 
     @Override
@@ -147,7 +136,7 @@ public class GetMailTask extends AsyncTask<Void, Void, MailBundle> {
     private Date getLastParseDate(String dateStr) {
         Date d;
         try {
-            d = dateFormat.parse(dateStr);
+            d = dateFormatter.parse(dateStr);
         } catch (ParseException e) {
             Calendar c = Calendar.getInstance();
             // Remember the first day of Ingress? Pepperidge Farm remembers.
